+++ EffectFramework.Core/Models/Db/AuditLog.cs
@@ -4,10 +4,10 @@ namespace EffectFramework.Core.Models.Db
 {
     public class AuditLog
     {
-        public int AuditLogID { get; set; }
-        public int ItemID { get; set; }
-        public int EntityID { get; set; }
-        public int? FieldID { get; set; }
+        public long AuditLogID { get; set; }
+        public long ItemID { get; set; }
+        public long EntityID { get; set; }
+        public long? FieldID { get; set; }
         public DateTime? EffectiveDateOld { get; set; }
         public DateTime? EffectiveDateNew { get; set; }
         public DateTime? EndEffectiveDateOld { get; set; }
@@ -20,18 +20,18 @@ namespace EffectFramework.Core.Models.Db
         public decimal? ValueDecimalNew { get; set; }
         public bool? ValueBooleanOld { get; set; }
         public bool? ValueBooleanNew { get; set; }
-        public int? ValueLookupOld { get; set; }
-        public int? ValueLookupNew { get; set; }
+        public long? ValueLookupOld { get; set; }
+        public long? ValueLookupNew { get; set; }
         public byte[] ValueBinaryOld { get; set; }
         public byte[] ValueBinaryNew { get; set; }
-        public int? ValueItemReferenceOld { get; set; }
-        public int? ValueItemReferenceNew { get; set; }
-        public int? ValueEntityReferenceOld { get; set; }
-        public int? ValueEntityReferenceNew { get; set; }
+        public long? ValueItemReferenceOld { get; set; }
+        public long? ValueItemReferenceNew { get; set; }
+        public long? ValueEntityReferenceOld { get; set; }
+        public long? ValueEntityReferenceNew { get; set; }
         public DateTime CreateDate { get; set; }
-        public int? ItemReference { get; set; }
+        public long? ItemReference { get; set; }
         public string Comment { get; set; }
-        public int TenantID { get; set; }
+        public long TenantID { get; set; }
 
     }
 }
