+++ EffectFramework.Core/Models/Db/ObjectIdentity.cs
@@ -5,7 +5,7 @@ namespace EffectFramework.Core.Models.Db
 {
     public class ObjectIdentity
     {
-        public int ObjectID;
+        public long ObjectID;
         public Guid ObjectGuid;
         public bool DidUpdate;
     }
